changed editor to edit
